// packages/api/src/routes/potions/types.ts
export type PotionsQuerystring = {
  id?: string;
  name?: string;
  effect?: string;
  side_effects?: string;
  characteristics?: string;
  time?: string;
  difficulty?: string;
  ingredients?: string;
  inventors?: string;
  manufacturers?: string;
  image?: string;
  wiki?: string;
  sort?: string;
  order?: string;
  page?: number;
  size?: number;
}